Refactor(InputView): 직접적인 ui를 담당하는 클래스들 패키지 변경
