+++ src/main/java/lotto/config/AppConfig.java
@@ -4,8 +4,8 @@ import lotto.application.LottoFactory;
 import lotto.application.LottoService;
 import lotto.application.RandomLottoFactory;
 import lotto.controller.LottoController;
-import lotto.view.InputView;
-import lotto.view.OutputView;
+import lotto.view.ui.InputView;
+import lotto.view.ui.OutputView;
 
 /**
  * 애플리케이션 객체의 의존 관계를 설정하고 조립하는 DI 컨테이너
